refactor: ImageController를 AdminController로 병합
