+++ api/src/main/kotlin/com/few/api/domain/admin/document/dto/PutImageUseCaseIn.kt
@@ -1,7 +1,7 @@
-package com.few.api.web.controller.image.request
+package com.few.api.domain.admin.document.dto
 
 import org.springframework.web.multipart.MultipartFile
 
-data class ImageSourceRequest(
+data class PutImageUseCaseIn(
     val source: MultipartFile
 )
