--- api/src/main/kotlin/com/few/api/domain/image/service/dto/GetUrlQuery.kt
@@ -1,5 +1,0 @@
-package com.few.api.domain.image.service.dto
-
-data class GetUrlQuery(
-    val `object`: String
-)
