+++ api/src/main/kotlin/com/few/api/web/controller/admin/request/ImageSourceRequest.kt
@@ -1,7 +1,7 @@
-package com.few.api.domain.image.dto
+package com.few.api.web.controller.admin.request
 
 import org.springframework.web.multipart.MultipartFile
 
-data class PutImageUseCaseIn(
+data class ImageSourceRequest(
     val source: MultipartFile
 )
